feat: add `fish_tmux_alter_path` config variable
The variable controles whether to change `PATH` automatically or not.
To preserve the existing behavior the variable default to true. This change
gives the user control over their `PATH`. For example, when using the `asdf`
version manager and homebrew in parallel, one would want to have `~/.asdf/bin`
folder before `/opt/homebrew/bin` on `PATH`.